refactor design
